+++ src/main/scala/org/constellation/infrastructure/p2p/PeerHealthCheckWatcher.scala
@@ -16,7 +16,7 @@ class PeerHealthCheckWatcher(
   unboundedHealthExecutionContext: ExecutionContext
 ) extends PeriodicIO("PeerHealthCheckWatcher", unboundedHealthExecutionContext: ExecutionContext) {
 
-  override def trigger(): IO[Unit] = healthCheckConsensusManager.checkPeersHealth()
+  override def trigger(): IO[Unit] = healthCheckConsensusManager.triggerHealthcheckManagement()
 
   schedule(30 seconds, ConfigUtil.getDurationFromConfig("constellation.health-check.p2p-interval", 5 seconds, config))
 }